Add files via upload
This is all the relevant code I put together for ACCDMG related analysis.  First part is on train speeds faster than the speed limit for the track class they were on.  I created a categorical variable called OverLimit which compared a quantitative variable called SpeedLimit to TRNSPD; if the train was over the speed limit, it was 1 else a 0.  The second part is ACCDMG vs. Weather, specifically precipitation or no precipitation (variable Precip set to 1 for rain, sleet, or snow otherwise 0).